rmFileAndDir
